server/requirements.txt
# Core Backend
fastapi
uvicorn
python-dotenv
requests

# Database
sqlmodel
psycopg2-binary

# Authentication
passlib[bcrypt]
bcrypt>=4.0.0,<5.0.0  # Pin to 4.x for passlib compatibility

# AI & Data Processing
google-generativeai
pdfplumber

# Optional / Legacy (Keep if needed for future ML)
numpy
pandas
joblib
scikit-learn
<<<<<<< HEAD
cryptography
qrcode
Pillow
=======
catboost
>>>>>>> 0776d26ec994cfa58aaca7a2d912300cbfe20d00
